Created a new file: minAreaCalculator.js that calculates the coordinates of the points that describe the shape of the polygon enclosed by the front and rear spars. These coordinates are output in two arrays: one for the x-coordinates and one for the y-coordinates.
I also created the quickVerificationPlot.py file to verify the coordinates that minAreaCalculator.js put out.
The quickVerificationPlot.py file is now useless, as there is nothing to be verified anymore.

Lastly, I changed the name of wingCGLocation.js to 1SPARwingCGLocation to clarify that the model is only to be
used for calculations in the case of 1 spar.